requirements.txt
# ─────────────────────────────────────────────
# Dermo-Scope – Python Dependencies
# ─────────────────────────────────────────────

# Core Deep Learning (TFLite – lightweight, cloud-compatible)
tflite-runtime

# Web Application
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# Computer Vision
opencv-python-headless>=4.8.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image Processing
Pillow>=10.0.0

# Media (required by streamlit-webrtc)
av>=10.0.0

# Optional – progress bars in training
tqdm>=4.65.0
